requirements.txt
torch==2.1.0
torchvision==0.16.0
segmentation-models-pytorch==0.3.3
albumentations==1.3.1
opencv-python==4.8.1.78
numpy==1.26.2
pandas==2.1.4
matplotlib==3.8.2
scikit-learn==1.3.2
pyyaml==6.0.1
tqdm==4.66.1
wandb==0.16.1
pytest==7.4.3
